Started sketching implementations of two-sided matching functions.
