+++ src/main/scala-2.11/markets/matching/onesided/FindFirstMatchingFunction.scala
@@ -17,32 +17,30 @@ package markets.matching
 
 import java.util.UUID
 
+import markets.matching.onesided.MatchingFunction
 import markets.orderbooks
-import markets.tradables.orders.{Operator, Order, Predicate}
+import markets.tradables.orders.{Order, Predicate}
 
 
-/** A `MatchingFunction` that matches an `Order` with its preferred match from a collection of acceptable matches.
+/** Class defining a `MatchingFunction` that finds the first acceptable `Order` in an `OrderBook`.
   *
   * @tparam O1 the type of `Order` instances that are potential matches.
   * @tparam O2 the type of `Order` that should be matched.
   */
-class FilterPreferredMatchingFunction[O1 <: Order, O2 <: Order with Predicate[O1] with Operator[O1]]
-  extends MatchingFunction[O1, O2] {
+class FindFirstMatchingFunction[O1 <: Order, O2 <: Order with Predicate[O1]] extends MatchingFunction[O1, O2] {
 
-  /** Matches an `Order` with its preferred match from a collection of acceptable matches.
+  /** Matches a given `Order` with the first acceptable `Order` found in some `OrderBook`.
     *
-    * @param order an `Order` in search of a match.
+    * @param order an `Order` of type `O2` in search of a match.
     * @param orderBook an `OrderBook` containing potential matches for the `order`.
-    * @return `None` if the `orderBook` is empty; `Some(order, matchingOrder)` otherwise.
+    * @return `None` if no acceptable `Order` is found in the `orderBook`; `Some(order, matchingOrder)` otherwise.
     * @note Worst case performance of this matching function is `O(n)` where `n` is the number of `Order` instances
-    *       contained in the `orderBook`.
+    *       contained in the `orderBook`.  Depending on the type of `orderBook`, the result of this `MatchingFunction`
+    *       may be non-deterministic.
     */
   def apply(order: O2, orderBook: orderbooks.OrderBook[O1, collection.GenMap[UUID, O1]]): Option[(O2, O1)] = {
-    orderBook.filter(order.isAcceptable) match {
-      case Some(orders) => orders.reduceOption(order.operator) match {
-        case Some(matchingOrder) => Some(order, matchingOrder)
-        case None => None
-      }
+    orderBook.find(order.isAcceptable) match {
+      case Some(matchingOrder) => Some(order, matchingOrder)
       case None => None
     }
   }
