+++ src/main/scala-2.11/markets/matching/onesided/MostPreferredMatchingFunction.scala
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package markets.matching
+package markets.matching.onesided
 
 import java.util.UUID
